Very Basic System
Has a tiny amount of code. Just enough to start us off.
Classes:
Line - Contains a String for its name and a linked list of <StationNode> objects
StationList - a list of all stations in the system, can create new StationNodes, and checks if a StationNode of that name has already been created. Used to reference all stations when creating a Line
StationNode  - a singular station in the system, has a name and can return it 
MtrGui - Empty so far